refactor(parakeet): share tokenizer json importer logic
